arm: mvebu: align device-tree with mainline linux
This commit does two things.  First it aligns the entries with
mainline linux.  Second it creates a minimal hardware profile
that can be used with both the ClearFog Pro and Base.  The only
limitation of this configuration is that only the eth0 port can
be used for tftp and pxe booting.  Even though I we only support
the single port we leave all 3 network interfaces in the
device-tree so we can generate and/or load MAC addresses for them.

Signed-off-by: Jon Nettleton <jon@solid-run.com>