imx: introduce CONFIG_IMX_SPL_BOOT_DEVICE_ menu option
This is a first stab at making an SOC specific boot menu
that makes enabling options for the various SPL boot methods
a little simpler.  There are many corner cases to be sorted
out however this gets the basic options selections done and
allows you to use the definition to tweak your machine config
according to what boot medium it is targeted at.

I will continue to work on the patchset until it is ready for
a RFC in mainline.

Signed-off-by: Jon Nettleton <jon@solid-run.com>