mx6cuboxi: set the ethaddr environment variable by default
When bringing up unfused soms in production or flexible
development environments it is best to have consistent MAC
addresses until the product is fused and ready for production.

This checks for the ethaddr environment variable, if it is not
set it checks the fuses first, and then generates a local MAC
address and saves the address that is found.  Once the fuse is
set this process can be reset by deleting the ethaddr environment
variable and rebooting.

Signed-off-by: Jon Nettleton <jon@solid-run.com>