mvebu: introduce CONFIG_MVEBU_SPL_BOOT_DEVICE_ menu option
This is an implementation of my imx based SOC specific boot menu
that makes enabling options for the various SPL boot methods
a little simpler.  There are many corner cases to be sorted
out however this gets the basic options selections done and
allows you to use the definition to tweak your machine config
according to what boot medium it is targeted at.

This adds support for the SolidRun Clearfog boards as well as
integrated the two independant menus that were being used for
the ControlcenterDC and Turris Omnia boards.

Signed-off-by: Jon Nettleton <jon@solid-run.com>