net: ping,arp: Fix cache alignment issues
Both ping_receive and arp_receive would transmit a received packet
back out using its original point.  This causes problesm with
certain network cards that add a custom header to the packet.
Specifically the mvneta driver for the Armada series boards has
a 2 byte Marvell header that is bypassed and passed along to
the system, but that 2 byte offset now causes a misalignment if
it is attempted to be sent back out.

Rather than changing the driver to memcpy all the received packets
to cache aligned buffers we instead change the two offending
network commands to copy the packet into a cache aligned net_tx_packet
before sending it back out.

Signed-off-by: Jon Nettleton <jon@solid-run.com>