mvebu: bubt: Properly support emmc hardware partitions
This commit does two fold.  First extends the mmc support
so that it will work with both DM and legacy devices.  Secondly
rather than using CONFIG_SYS_MMC_ENV_PART to determine where
to flash the image we instead query the eMMC to see which
hardware partition is selected to boot from.  It is very possible
that you would want to boot from one hardware partition and store
the environment somewhere completely different.

If you want to flash to a different boot partition you must first
use the mmc partconf command to activate the partition you want
to write to.

Signed-off-by: Jon Nettleton <jon@solid-run.com>