rtc: mvebu: Detect if RTC is in an unintialized state
It is not definitive to know if the RTC is in an uninialized state or not,
but the following call will read some bits in the RTC unit and guess if
if it's in that state, and accordingly set it to sane default values.

This catches RTC misfunctions that can keep systemd from booting.

Signed-off-by: Jon Nettleton <jon@solid-run.com>