net: core: Allow incremental MAC addresses
Rather than generating multiple completely different random MAC
addresses for multiple interfaces, we try to use the first
randomly generated MAC address and then increment it for the
reset of the interfaces that need addresses.

enabled with NET_SEQ_MACADDR_FROM_ENV

Additionally we add a config option to designate if the randomly
generated MACs should be added to the environment.

enabled with NET_RANDOM_ETHADDR_TO_ENV

Signed-off-by: Jon Nettleton <jon@solid-run.com>