mvebu: env: set a sane env device based on boot mode
This is some Kconfig trickery that probably needs more work.
We set a bool kconfig option ENV_IS_ON_SELECTED_BOOT_DEVICE
that disables the env mode until a choice from the mvebu
boot menu is chosen.  This allows a defconfig to be initialized
and then using menuconfig you can enable the boot mode of choice
and it will default to the appropriate environment storage
location.  All other architectures will default to y for
ENV_IS_ON_SELECTED_BOOT_DEVICE so the environment location
choice will work as it used to.

Signed-off-by: Jon Nettleton <jon@solid-run.com>