mvebu: clearfog: unify device environment setup
This sets up all devices to save the environment
of 8K in size at a negative offset of the 1MB location
on the specific medium.  This fits before the default
initial partion location of 1MB for sdhc cards, and
leaves plenty of room on both our SPI flash and the eMMC
hardware boot partitions.  There is also enough empty
space so if u-boot enlarges it won't overwrite the existing
environment.

Additionally remove the default environment from defconfig

Signed-off-by: Jon Nettleton <jon@solid-run.com>