mvebu: clearfog: Enable MAC sanity for the Clearfog
The A38x does not have enough on board fuses for both
secure boot keys and the MAC addresses of the on board devices.
In order to try and make the device more usable in a dev
environment we do the following things.

1) Generate an initial MAC based on the SolidRun OUI
2) Generate the additional 2 MACs as incremental address
3) expose all the addresses to the environment
4) use the address to populate the device-tree before loading linux.

We are not saving the environment by default so to always have
the same MAC addresses you should save the env after the
first boot.

Signed-off-by: Jon Nettleton <jon@solid-run.com>