mvebu: disable pl310-cache double-linefill
As per commit cda80a82ac3e89309706c027ada6ab232be1d640 in the
Linux mainline kernel the double-linefill functionality has
been found to cause issues.  This is unlikely to happen under
u-boot however we will disable it here regardless.

This should most likely be moved to a cache initialization
function, however the cache is already disabled here, so we may
as well update the bits.

Signed-off-by: Jon Nettleton <jon@solid-run.com>